задание к спринту 3 - add tests for functional section
