+++ functional/src/main/kotlin/ru/sber/functional/PowFactory.kt
@@ -8,11 +8,3 @@ object PowFactory {
      */
     fun buildPowFunction(pow: Int) = { n: Number -> n.toDouble().pow(pow) }
 }
-
-fun main() {
-    val powFunc = PowFactory.buildPowFunction(10)
-
-    println(powFunc(2L))
-    println(powFunc(3))
-    println(PowFactory.buildPowFunction(2)(3))
-}
